exceptions

+++ src/main/java/com/example/taskmanagement/user/service/UserService.java
@@ -1,5 +1,8 @@
 package com.example.taskmanagement.user.service;
 
+import com.example.taskmanagement.exceptions.UserHasTasksException;
+import com.example.taskmanagement.exceptions.UserNotFoundException;
+import com.example.taskmanagement.task.repository.TaskRepository;
 import com.example.taskmanagement.user.domain.UserEntity;
 import com.example.taskmanagement.user.dto.UserDto;
 import com.example.taskmanagement.user.mapper.UserMapper;
@@ -13,6 +16,7 @@ import org.springframework.stereotype.Service;
 public class UserService {
 
   private final UserRepository userRepository;
+  private final TaskRepository taskRepository;
   private final UserMapper userMapper;
 
   public List<UserDto> getAllUsers() {
@@ -28,10 +32,13 @@ public class UserService {
     return userRepository
         .findById(id)
         .map(userMapper::toDto)
-        .orElseThrow(() -> new RuntimeException("User not found"));
+        .orElseThrow(() -> new UserNotFoundException("User not found"));
   }
 
   public void deleteUserById(Long id) {
+    if (taskRepository.existsByAssigneeId(id)) {
+      throw new UserHasTasksException("User can not be deleted. Has tasks assigned.");
+    }
     userRepository.deleteById(id);
   }
 }
